UTF-8 as character encoding for SLDs
This is the following of the previous incomplete PR11, generalizing
switch to UTF-8 for all styles.

Having the styles in ISO messes up things when an user edits them via
the UI editor from GS (as the UI is supposed to be full utf-8,
everything posted will be interpreted as UTF-8, not ISO, messing up the
charset).

Note: if we git grep onto 'ISO-8859' in the repository, there are still
some references, but on automated_tests layer ; a shapefile is provided
where the attributes are still encoded in this charset.